refactor 전반적인 예외처리 관리
