+++ src/main/java/com/grepp/synapse4/app/controller/api/restaurant/MainSearchApiController.java
@@ -3,10 +3,13 @@ package com.grepp.synapse4.app.controller.api.restaurant;
 import com.grepp.synapse4.app.model.restaurant.RestaurantSearchService;
 import com.grepp.synapse4.app.model.restaurant.dto.search.SearchRestaurantRequestDto;
 import com.grepp.synapse4.app.model.restaurant.dto.search.SearchRestaurantResponseDto;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -14,15 +17,17 @@ import java.util.List;
 @RequestMapping
 public class MainSearchApiController {
 
-    // todo result view fetch 처리... 하고 싶다.
     private final RestaurantSearchService restaurantSearchService;
 
     @GetMapping("/api/search")
-    public ResponseEntity<List<SearchRestaurantResponseDto>> searchRestaurants(
-            @ModelAttribute SearchRestaurantRequestDto requestDto) {
+    public ResponseEntity<?> searchRestaurants(@Valid @ModelAttribute SearchRestaurantRequestDto requestDto,
+                                               BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            String msg = bindingResult.getFieldError().getDefaultMessage();
+            return ResponseEntity.badRequest().body(msg); // -> "검색어는 필수입니다."
+        }
 
         List<SearchRestaurantResponseDto> respon = restaurantSearchService.searchByName(requestDto.getRestaurantKeyword());
         return ResponseEntity.ok(respon);
     }
-
 }
