+++ src/main/java/com/grepp/synapse4/app/controller/web/recommend/CurationController.java
@@ -20,6 +20,15 @@ public class CurationController {
     public String curationList(Model model) {
         CurationDto curation =
                 curationService.getLatestCurationRestaurants();
+
+        // DB 에 적절한 데이터가 없을 때 띄어주는 화면이 있으면 여러 컨트롤러에서
+        // 걸리는 같은 이슈를 묶어 담을 수 있을 듯
+        // 일단은 뷰에서 분기 처리..
+//        if(curation == null){
+//            model.addAttribute("message", "큐레이션 결과가 없습니다.");
+//            return "noList";
+//        }
+
         model.addAttribute("curation", curation);
 
         return "recommend/userCuration";
