+++ src/main/java/com/grepp/synapse4/app/model/llm/dto/GeminiUserRequestDto.java
@@ -1,5 +1,6 @@
 package com.grepp.synapse4.app.model.llm.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,10 +8,10 @@ import lombok.Setter;
 import org.springframework.stereotype.Service;
 
 @Getter
-@Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class GeminiUserRequestDto {
 
+    @NotBlank(message = "입력값은 필수입니다.")
     private String userText;
 }
