+++ src/main/java/com/grepp/synapse4/app/model/restaurant/dto/search/SearchRestaurantRequestDto.java
@@ -8,6 +8,6 @@ import lombok.*;
 @AllArgsConstructor
 public class SearchRestaurantRequestDto {
 
-    @NotBlank
+    @NotBlank(message = "입력값은 필수입니다.")
     private String restaurantKeyword;
 }
